Remove JSON schema validation
We no longer need to do schema validation in the DLME web application because this is happening in the dlme-transform component, which happens earlier in the data flow (before the data gets loaded into the web app). This means we no longer need to keep our schemas in two places and try to keep them in sync. It also means we will be ditching the ability to upload DLME JSON into the app via a textarea form field; data will only flow into the application from the dlme-transform staging area.

Fixes #597